+++ file/file.go
@@ -1,8 +1,3 @@
-// This module expects the following environment variables:
-//
-// PROXYM_FILE_CONFIGS_PATH - Absolute path to the location of the JSON configuration files.
-//
-// PROXYM_FILE_ENABLED - Enable this module.
 package file
 
 import (