+++ haproxy/haproxy.go
@@ -1,18 +1,3 @@
-// A ConfigGenerator that generates a configuration file for HAProxy.
-// Restarts HAProxy with the new config in case there are changes.
-// It uses domain based routing for HTTP applications.
-//
-// It requires the following environment variables to be set:
-// PROXYM_HAPROXY_BINARY_PATH - The absolute path to the binary of HAProxy
-//
-// PROXYM_HAPROXY_CONFIG_FILE_PATH - An absolute path where the generated config file will be stored.
-//
-// PROXYM_HAPROXY_ENABLED - Enable this module.
-//
-// PROXYM_HAPROXY_OPTIONS_PATH - An absolute path to a file where 'global' options and 'default's of HAProxy are stored. The
-//                               content of this file prepended to the rest of the config.
-//
-// PROXYM_HAPROXY_PID_PATH - An absolute path where HAProxy stores its PID.
 package haproxy
 
 import (
@@ -23,19 +8,48 @@ import (
 	"github.com/wndhydrnt/proxym/log"
 	"github.com/wndhydrnt/proxym/manager"
 	"github.com/wndhydrnt/proxym/types"
+	"gopkg.in/yaml.v2"
 	"io/ioutil"
 	"os"
 	"os/exec"
 	"strings"
+	"text/template"
 )
 
 type Config struct {
-	BinaryPath     string `envconfig:"binary_path"`
-	ConfigFilePath string `envconfig:"config_file_path"`
-	Enabled        bool
-	HttpPort       int    `envconfig:"http_port"`
-	OptionsPath    string `envconfig:"options_path"`
-	PidPath        string `envconfig:"pid_path"`
+	BinaryPath       string `envconfig:"binary_path"`
+	ConfigFilePath   string `envconfig:"config_file_path"`
+	Enabled          bool
+	PidPath          string `envconfig:"pid_path"`
+	SettingsPath     string `envconfig:"settings_path"`
+	TemplateFilePath string `envconfig:"template_file_path"`
+}
+
+type ServiceAndSettings struct {
+	Settings *ServiceSettings
+	Service  types.Service
+}
+
+type ServiceSettings struct {
+	Domains  []string
+	Config   string
+	Protocol string
+}
+
+func (s *ServiceSettings) ConfigItems() []string {
+	items := []string{}
+
+	lines := strings.Split(s.Config, "\n")
+	for _, line := range lines {
+		line = strings.Trim(line, " ")
+		if line == "" {
+			continue
+		}
+
+		items = append(items, line)
+	}
+
+	return items
 }
 
 type HAProxyGenerator struct {
@@ -89,230 +103,69 @@ func (h *HAProxyGenerator) Generate(services []types.Service) {
 }
 
 func (h *HAProxyGenerator) config(services []types.Service) string {
-	var httpServices []types.Service
-	var tcpServices []types.Service
+	var srvcs []ServiceAndSettings
 
 	for _, service := range services {
+		serviceConfig, _ := h.readServiceSettings(service.Domain, service.Port)
 
-		haproxyProtocol, _ := h.readProtocol(service.Domain, service.Port)
-
-		if service.Protocol == "tcp" && strings.TrimSpace(haproxyProtocol) == "http" {
-			httpServices = append(httpServices, service)
+		if service.Protocol != "tcp" {
+			// HAProxy supports TCP only. Ignore any other protocol.
 			continue
 		}
 
-		if service.Protocol == "tcp" {
-			tcpServices = append(tcpServices, service)
-		}
-		// HAProxy supports TCP only. Ignore any other protocol.
+		srvcs = append(srvcs, ServiceAndSettings{serviceConfig, service})
 	}
 
-	globalConfig, err := readExistingFile(h.c.OptionsPath + "/global.cfg")
+	globalConfig, err := readExistingFile(h.c.TemplateFilePath)
 	if err != nil {
 		log.ErrorLog.Error("Unable to read global config. Stopping HAProxy config generator: %s", err)
 		return ""
 	}
 
-	httpConfig := h.httpConfig(httpServices)
-	tcpConfig := h.tcpConfig(tcpServices)
-
-	fullConfig := globalConfig + httpConfig + tcpConfig
-
-	return fullConfig
-}
-
-func (h *HAProxyGenerator) httpConfig(services []types.Service) string {
-	var buffer bytes.Buffer
-	aclLines := []string{}
-	useBackendLines := []string{}
-
-	if len(services) == 0 {
+	tlt, err := template.New("haproxy").Parse(globalConfig)
+	if err != nil {
+		log.ErrorLog.Error("%s", err)
 		return ""
 	}
 
-	buffer.WriteString("\nfrontend http-in\n  bind *:80\n\n")
+	var out bytes.Buffer
 
-	for _, service := range services {
-		name := h.generateName(service.Id)
-
-		aclLine := fmt.Sprintf("  acl host_%s hdr(host) -i %s", name, service.Domain)
-		aclLines = append(aclLines, aclLine)
-
-		additionalDomains, err := h.readDomains(service.Domain, service.Port)
-		if err != nil {
-			log.ErrorLog.Error("Error reading domains file: '%s'", err)
-		}
-
-		for _, d := range additionalDomains {
-			aclLines = append(aclLines, fmt.Sprintf("  acl host_%s hdr(host) -i %s", name, d))
-		}
-
-		useBackendLine := fmt.Sprintf("  use_backend %s_cluster if host_%s", name, name)
-		useBackendLines = append(useBackendLines, useBackendLine)
-	}
-
-	buffer.WriteString(strings.Join(aclLines, "\n"))
-	buffer.WriteString("\n\n")
-	buffer.WriteString(strings.Join(useBackendLines, "\n"))
-	buffer.WriteString("\n")
-
-	for _, service := range services {
-		buffer.WriteString("\n")
-
-		name := h.generateName(service.Id)
-
-		buffer.WriteString(fmt.Sprintf("backend %s_cluster\n", name))
-
-		optionData, err := h.readConfig(service.Domain, service.Port)
-		if err != nil {
-			log.ErrorLog.Error("Error reading options file: '%s'", err)
-		}
-
-		for _, option := range h.sanitizeOptions(optionData) {
-			buffer.WriteString(fmt.Sprintf("  %s\n", option))
-		}
-
-		buffer.WriteString("\n")
-
-		for i, host := range service.Hosts {
-			nodeNumber := i + 1
-			buffer.WriteString(fmt.Sprintf("  server node%d %s:%d check\n", nodeNumber, host.Ip, host.Port))
-		}
-	}
-
-	return buffer.String()
-}
-
-func (h *HAProxyGenerator) tcpConfig(services []types.Service) string {
-	var buffer bytes.Buffer
-
-	for _, service := range services {
-		name := h.generateName(service.Id)
-		header := fmt.Sprintf("\nlisten %s :%d\n  mode tcp\n", name, service.ListenPort())
-		buffer.WriteString(header)
-
-		optionData, err := h.readConfig(service.Domain, service.Port)
-		if err != nil {
-			log.ErrorLog.Error("Error reading options file: '%s'", err)
-		}
-
-		for _, option := range h.sanitizeOptions(optionData) {
-			buffer.WriteString(fmt.Sprintf("  %s\n", option))
-		}
-
-		buffer.WriteString("\n")
-
-		for index, host := range service.Hosts {
-			nodeNumber := index + 1
-			serverLine := fmt.Sprintf("  server node%d %s:%d check\n", nodeNumber, host.Ip, host.Port)
-			buffer.WriteString(serverLine)
-		}
-	}
-
-	return buffer.String()
-}
-
-func (h *HAProxyGenerator) generateName(id string) string {
-	if strings.Contains(id, "/") {
-		parts := strings.Split(id, "/")
-
-		// Remove empty part in case of leading '/' in id
-		if parts[0] == "" {
-			parts = parts[1:]
-		}
-
-		return strings.Join(parts, "_")
-	}
-
-	return id
-}
-
-func (h *HAProxyGenerator) sanitizeOptions(o string) []string {
-	sanitizedOptions := []string{}
-
-	options := strings.Split(o, "\n")
-	for _, option := range options {
-		option = strings.Trim(option, " ")
-		if option == "" {
-			continue
-		}
-
-		sanitizedOptions = append(sanitizedOptions, option)
-	}
-
-	return sanitizedOptions
-}
-
-func (h *HAProxyGenerator) readConfig(domain string, port int) (string, error) {
-	optionPath := fmt.Sprintf("%s/%s_%d/config", h.c.OptionsPath, domain, port)
-
-	_, err := os.Stat(optionPath)
+	err = tlt.Execute(&out, srvcs)
 	if err != nil {
-		return "", nil
+		log.ErrorLog.Error("%s", err)
+		return ""
+	}
+
+	return removeEmptyLines(out.String()) + "\n"
+}
+
+func (h *HAProxyGenerator) readServiceSettings(domain string, port int) (*ServiceSettings, error) {
+	path := fmt.Sprintf("%s/%s_%d.yml", h.c.SettingsPath, domain, port)
+
+	_, err := os.Stat(path)
+	if err != nil {
+		return &ServiceSettings{}, nil
 	} else {
-		f, err := os.Open(optionPath)
+		f, err := os.Open(path)
 		if err != nil {
-			return "", err
+			return &ServiceSettings{}, err
 		}
 
 		defer f.Close()
 
-		options, err := ioutil.ReadAll(f)
+		data, err := ioutil.ReadAll(f)
 		if err != nil {
-			return "", err
+			return &ServiceSettings{}, err
 		}
 
-		return string(options), nil
-	}
-}
+		serviceConfig := &ServiceSettings{}
 
-func (h *HAProxyGenerator) readDomains(domain string, port int) ([]string, error) {
-	domains := []string{}
-
-	domainsPath := fmt.Sprintf("%s/%s_%d/domains", h.c.OptionsPath, domain, port)
-
-	_, err := os.Stat(domainsPath)
-	if err != nil {
-		return domains, nil
-	} else {
-		f, err := os.Open(domainsPath)
+		err = yaml.Unmarshal(data, serviceConfig)
 		if err != nil {
-			return domains, err
+			return &ServiceSettings{}, err
 		}
 
-		defer f.Close()
-
-		content, err := ioutil.ReadAll(f)
-		if err != nil {
-			return domains, err
-		}
-
-		sanitizedContent := strings.Trim(string(content), "\n")
-
-		return strings.Split(sanitizedContent, "\n"), nil
-	}
-}
-
-func (h *HAProxyGenerator) readProtocol(domain string, port int) (string, error) {
-	protocolPath := fmt.Sprintf("%s/%s_%d/protocol", h.c.OptionsPath, domain, port)
-
-	_, err := os.Stat(protocolPath)
-	if err != nil {
-		return "", nil
-	} else {
-		f, err := os.Open(protocolPath)
-		if err != nil {
-			return "", err
-		}
-
-		defer f.Close()
-
-		protocol, err := ioutil.ReadAll(f)
-		if err != nil {
-			return "", err
-		}
-
-		return string(protocol), nil
+		return serviceConfig, nil
 	}
 }
 
@@ -334,6 +187,21 @@ func readExistingFile(fp string) (string, error) {
 	}
 
 	return string(content), nil
+}
+
+// Helper that removes all lines in a string that consist only of spaces.
+func removeEmptyLines(in string) string {
+	newLines := []string{}
+
+	lines := strings.Split(in, "\n")
+
+	for _, line := range lines {
+		if strings.TrimSpace(line) != "" {
+			newLines = append(newLines, line)
+		}
+	}
+
+	return strings.Join(newLines, "\n")
 }
 
 // Create a new HAProxyGenerator, reading configuration from environment variables.